Refactor: separate models into distinct files for User and Region.
